requirements.txt
asgiref==3.9.1
boto3==1.40.39
botocore==1.40.39
dj-database-url==3.0.1
Django==5.2
django-storages==1.14.6
gunicorn==23.0.0
jmespath==1.0.1
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0
